predicta_backend_dj42/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
pdfminer.six>=202312
python-docx>=1.1
Pillow>=10.0
python-dotenv>=1.0
psycopg2-binary>=2.9
whitenoise>=6.7
requests>=2.32.0

